+++ src/Viz.js
@@ -124,6 +124,15 @@ export default class Viz {
         this._validateAliasDAG();
     }
 
+    /**
+     * Get the geometry type of the visualization.
+     * @readonly
+     * @api
+     */
+    get geometryType () {
+        return this._geomType;
+    }
+
     loadImages () {
         return Promise.all(this._getRootExpressions().map(expr => expr.loadImages()));
     }
@@ -506,6 +515,29 @@ export default class Viz {
                 console.warn(`Property '${key}' is not supported`);
             }
         }
+    }
+
+    /**
+     * Stringify the visualization
+     *
+     * @returns {string}
+     * @memberof carto
+     * @api
+     */
+    toString () {
+        const variables = Object.keys(this.variables).map(varName =>
+            `@${varName}: ${this.variables[varName].toString()}\n`
+        );
+        return `color: ${this.color.toString()}
+            strokeColor: ${this.strokeColor.toString()}
+            width: ${this.width.toString()}
+            strokeWidth: ${this.strokeWidth.toString()}
+            filter: ${this.filter.toString()}
+            order: ${this.order.toString()}
+            symbol: ${this.symbol.toString()}
+            symbolPlacement: ${this.symbolPlacement.toString()}
+            offset: ${this.offset.toString()}
+            ${variables}`.replace(/ {4}/g, '');
     }
 }